Implement DOI feature (#24)
* Implement convenience feature for users to submit result files to Zenodo and obtain a DOI
* Prereserve DOIs and put them into netcdf files before publishing
* Add user attribute ORCID
* Switched to new domain qa4sm.eu
* Prevent deletion of DOI published results
* Add published results page
* Introduce published validations button on landing page and in title bar, move menu items to dropdown menu
* Switched to different import method for django settings
* Introduce access token for zenodo into automatic builds
* Add documentation on Zenodo access tokens
* Allow editing validation runs in admin ui without setting filters or output files
* Remove requirement to be logged in for viewing results
* Show publication metadata to user for correction before publishing
* Updating online help with publication process and fixing admin email

+++ validator/views/results.py
@@ -5,8 +5,9 @@ from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.http import QueryDict
 from django.http.response import HttpResponse
 from django.shortcuts import get_object_or_404, render
-from django.utils import timezone
 
+from validator.doi import get_doi_for_validation
+from validator.forms import PublishingForm
 from validator.models import ValidationRun
 from validator.validation.globals import METRICS
 from validator.validation.graphics import get_dataset_pairs
@@ -33,7 +34,6 @@ def user_runs(request):
     return render(request, 'validator/user_runs.html', context)
 
 
-@login_required(login_url='/login/')
 def result(request, result_uuid):
     val_run = get_object_or_404(ValidationRun, pk=result_uuid)
 
@@ -41,6 +41,10 @@ def result(request, result_uuid):
         ## make sure only the owner of a validation can delete it (others are allowed to GET it, though)
         if(val_run.user != request.user):
             return HttpResponse(status=403)
+
+        ## check that our validation can be deleted; it can't if it already has a DOI
+        if(not val_run.is_unpublished):
+            return HttpResponse(status=405)
 
         val_run.delete()
         return HttpResponse("Deleted.", status=200)
@@ -56,7 +60,7 @@ def result(request, result_uuid):
             archive_mode = patch_params['archive']
 
             if not ((archive_mode == 'true') or (archive_mode == 'false')):
-                return HttpResponse("Wrong parameter.", status=400)
+                return HttpResponse("Wrong action parameter.", status=400)
 
             val_run.archive(unarchive = (archive_mode == 'false'))
             return HttpResponse("Changed.", status=200)
@@ -65,12 +69,33 @@ def result(request, result_uuid):
             extend = patch_params['extend']
 
             if extend != 'true':
-                return HttpResponse("Wrong parameter.", status=400)
+                return HttpResponse("Wrong action parameter.", status=400)
 
             val_run.extend_lifespan()
             return HttpResponse(val_run.expiry_date, status=200)
 
-        return HttpResponse("Wrong parameter.", status=400)
+        if 'publish' in patch_params:
+            publish = patch_params['publish']
+
+            # check we've got the action set correctly
+            if publish != 'true':
+                return HttpResponse("Wrong action parameter.", status=400)
+
+            # check that the publication parameters are valid
+            pub_form = PublishingForm(data=patch_params, validation=val_run)
+            if not pub_form.is_valid():
+                # if not, send back an updated publication form with errors set and http code 420 (picked up in javascript)
+                return render(request, 'validator/publishing_dialog.html', {'publishing_form': pub_form, 'val': val_run}, status=420)
+
+            try:
+                get_doi_for_validation(val_run, pub_form.pub_metadata)
+            except Exception as e:
+                m = getattr(e, 'message', repr(e))
+                return HttpResponse(m, status=400)
+
+            return HttpResponse("Published.", status=200)
+
+        return HttpResponse("Wrong action parameter.", status=400)
 
     # by default, show page
     else:
@@ -89,6 +114,9 @@ def result(request, result_uuid):
 
         pairs = get_dataset_pairs(val_run)
 
+        # the publication form is only needed by the owner; if we're displaying for another user, avoid leaking user data
+        pub_form = PublishingForm(validation=val_run) if is_owner else None
+
         context = {
             'is_owner': is_owner,
             'val' : val_run,
@@ -97,6 +125,7 @@ def result(request, result_uuid):
             'metrics': METRICS,
             'pairs': pairs,
             'json_metrics': json_dumps(METRICS),
+            'publishing_form': pub_form
             }
 
         return render(request, 'validator/result.html', context)
